pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alimentador-bd"
version = "1.0.0"
description = "OLTP Hospital Simulator - Continuous data streaming for CDC testing with Debezium"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Henrique Ferreira", email = "henrique@example.com"}
]
keywords = [
    "postgresql",
    "cdc",
    "change-data-capture",
    "debezium",
    "oltp",
    "streaming",
    "testing",
    "data-generation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Portuguese (Brazilian)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Monitoring"
]

dependencies = [
    "psycopg2-binary>=2.9.9,<3.0",
    "python-dotenv>=1.0.0,<2.0",
    "typer>=0.12.0,<1.0",
    "faker>=21.0.0,<22.0",
    "pydantic>=2.0.0,<3.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "types-psycopg2>=2.9.21.0"
]

[project.urls]
Homepage = "https://github.com/yourusername/alimentador-bd"
Documentation = "https://github.com/yourusername/alimentador-bd#readme"
Repository = "https://github.com/yourusername/alimentador-bd.git"
Issues = "https://github.com/yourusername/alimentador-bd/issues"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "W503",  # line break before binary operator
]

[tool.ruff.isort]
profile = "black"
multi_line_mode = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "faker.*",
    "psycopg2.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --cov=scripts --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["scripts"]
omit = [
    "*/tests/*",
    "*/__main__.py",
    "*/venv/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*Protocol.*:",
    "@(abc\\.)?abstractmethod",
]
